Provide consistency in supervisor clap commands, correcting several help menu problems
Resolves https://github.com/habitat-sh/habitat/issues/5363
Resolves https://github.com/habitat-sh/habitat/issues/5377

This PR does several things, enumerated below:

- Moves all supervisor related clap subcommand definitions into functions
in hab::cli so that they are consistent and not duplicated in `hab-sup`
where they can be prone to drift.
- Enumerates _all_ supervisor related clap subcommands under the `hab-sup`
clap app, including the supervisor commands only implemented in `hab`.
This is so that the _full_ supervisor help menu can be displayed for
all supervisor related commands in _both_ `hab` and `hab-sup`.
Previously, depending on which `sup * --help` command
was run, different help menus would be displayed. This was the cause of
- Delegate `hab help sup` to `hab-sup`, passing only "help", so that the
full help menu was displayed. Previously, this command would route to `hab`,
displaying only the supervisor related commands run from `hab`
- Defines a custom usage message for all `hab-sup` related supervisor
commands, making it clearer that the correct usage is `hab sup ..` not
`hab-sup ..`
- Delegates `hab sup run -h` and `hab sup run --help` to `hab-sup` where
clap then correctly displays the help menu as opposed to routing to the
launcher where the launcher would actually attempt a start. This was
the cause of #5377 above.

Demonstrating the correct full help menu of all supervisor related
commands in `hab` and `hab-sup`:

```
✔ ~/habitat [jeremymv2/dry_up_sup_clap L|●5✚ 2⚑ 14]
17:28 $ sudo -E RUST_LOG=info target/debug/hab sup --help
hab-sup 0.60.0-dev

Authors: The Habitat Maintainers <humans@habitat.sh>
The Habitat Supervisor

USAGE:
    hab sup <SUBCOMMAND>

FLAGS:
    -h, --help       Prints help information
    -V, --version    Prints version information

SUBCOMMANDS:
    bash      Start an interactive Bash-like shell
    depart    Depart a Supervisor from the gossip ring; kicking and banning the target from joining again with the
              same member-id
    help      Prints this message or the help of the given subcommand(s)
    run       Run the Habitat Supervisor
    secret    Commands relating to a Habitat Supervisor's Contorl Gateway secret
    sh        Start an interactive Bourne-like shell
    status    Query the status of Habitat services.
    term      Gracefully terminate the Habitat Supervisor and all of its running services
✔ ~/habitat [jeremymv2/dry_up_sup_clap L|●5✚ 2⚑ 14]
17:28 $
```
